ironic-inspector 3.2.0 release

meta:version: 3.2.0
meta:series: mitaka
meta:release-type: release
meta:announce: openstack-announce@lists.openstack.org
meta:pypi: no
meta:first: no

+++ devstack/plugin.sh
@@ -23,13 +23,16 @@ IRONIC_INSPECTOR_RAMDISK_FLAVOR=${IRONIC_INSPECTOR_RAMDISK_FLAVOR:-fedora $IRONI
 IRONIC_INSPECTOR_COLLECTORS=${IRONIC_INSPECTOR_COLLECTORS:-default,logs}
 IRONIC_INSPECTOR_RAMDISK_LOGDIR=${IRONIC_INSPECTOR_RAMDISK_LOGDIR:-$IRONIC_INSPECTOR_DATA_DIR/ramdisk-logs}
 IRONIC_INSPECTOR_ALWAYS_STORE_RAMDISK_LOGS=${IRONIC_INSPECTOR_ALWAYS_STORE_RAMDISK_LOGS:-True}
+IRONIC_INSPECTOR_TIMEOUT=${IRONIC_INSPECTOR_TIMEOUT:-600}
 # These should not overlap with other ranges/networks
 IRONIC_INSPECTOR_INTERNAL_IP=${IRONIC_INSPECTOR_INTERNAL_IP:-172.24.42.254}
 IRONIC_INSPECTOR_INTERNAL_SUBNET_SIZE=${IRONIC_INSPECTOR_INTERNAL_SUBNET_SIZE:-24}
 IRONIC_INSPECTOR_DHCP_RANGE=${IRONIC_INSPECTOR_DHCP_RANGE:-172.24.42.100,172.24.42.253}
 IRONIC_INSPECTOR_INTERFACE=${IRONIC_INSPECTOR_INTERFACE:-br-inspector}
 IRONIC_INSPECTOR_INTERNAL_URI="http://$IRONIC_INSPECTOR_INTERNAL_IP:$IRONIC_INSPECTOR_PORT"
-IRONIC_INSPECTOR_INTERNAL_IP_WITH_NET=$IRONIC_INSPECTOR_INTERNAL_IP/$IRONIC_INSPECTOR_INTERNAL_SUBNET_SIZE
+IRONIC_INSPECTOR_INTERNAL_IP_WITH_NET="$IRONIC_INSPECTOR_INTERNAL_IP/$IRONIC_INSPECTOR_INTERNAL_SUBNET_SIZE"
+
+IRONIC_INSPECTOR_NODE_NOT_FOUND_HOOK=${IRONIC_INSPECTOR_NODE_NOT_FOUND_HOOK:-""}
 
 GITDIR["python-ironic-inspector-client"]=$DEST/python-ironic-inspector-client
 GITREPO["python-ironic-inspector-client"]=${IRONIC_INSPECTOR_CLIENT_REPO:-${GIT_BASE}/openstack/python-ironic-inspector-client.git}
@@ -97,30 +100,32 @@ function inspector_uses_ipa {
 
 function prepare_tftp {
     IRONIC_INSPECTOR_IMAGE_PATH="$TOP_DIR/files/ironic-inspector"
+    IRONIC_INSPECTOR_KERNEL_PATH="$IRONIC_INSPECTOR_IMAGE_PATH.kernel"
     IRONIC_INSPECTOR_INITRAMFS_PATH="$IRONIC_INSPECTOR_IMAGE_PATH.initramfs"
     IRONIC_INSPECTOR_CALLBACK_URI="$IRONIC_INSPECTOR_INTERNAL_URI/v1/continue"
 
-    if  inspector_uses_ipa; then
+    if inspector_uses_ipa; then
         IRONIC_INSPECTOR_KERNEL_CMDLINE="ipa-inspection-callback-url=$IRONIC_INSPECTOR_CALLBACK_URI systemd.journald.forward_to_console=yes"
+        IRONIC_INSPECTOR_KERNEL_CMDLINE="$IRONIC_INSPECTOR_KERNEL_CMDLINE vga=normal console=tty0 console=ttyS0"
         IRONIC_INSPECTOR_KERNEL_CMDLINE="$IRONIC_INSPECTOR_KERNEL_CMDLINE ipa-inspection-collectors=$IRONIC_INSPECTOR_COLLECTORS"
+        IRONIC_INSPECTOR_KERNEL_CMDLINE="$IRONIC_INSPECTOR_KERNEL_CMDLINE ipa-debug=1"
         if [[ "$IRONIC_INSPECTOR_BUILD_RAMDISK" == "True" ]]; then
-            IRONIC_INSPECTOR_KERNEL_PATH="$IRONIC_INSPECTOR_IMAGE_PATH.vmlinuz"
             if [ ! -e "$IRONIC_INSPECTOR_KERNEL_PATH" -o ! -e "$IRONIC_INSPECTOR_INITRAMFS_PATH" ]; then
-                if [[ $(type -P disk-image-create) == "" ]]; then
-                    pip_install_gr diskimage-builder
-                fi
-                disk-image-create $IRONIC_INSPECTOR_RAMDISK_FLAVOR \
-                    -o $IRONIC_INSPECTOR_IMAGE_PATH
-                sudo chown $STACK_USER $IRONIC_INSPECTOR_KERNEL_PATH
+                build_ipa_coreos_ramdisk "$IRONIC_INSPECTOR_KERNEL_PATH" "$IRONIC_INSPECTOR_INITRAMFS_PATH"
             fi
         else
             # download the agent image tarball
-            IRONIC_INSPECTOR_KERNEL_PATH="$IRONIC_INSPECTOR_IMAGE_PATH.kernel"
-            wget "$IRONIC_AGENT_KERNEL_URL" -O $IRONIC_INSPECTOR_KERNEL_PATH
-            wget "$IRONIC_AGENT_RAMDISK_URL" -O $IRONIC_INSPECTOR_INITRAMFS_PATH
+            if [ ! -e "$IRONIC_INSPECTOR_KERNEL_PATH" -o ! -e "$IRONIC_INSPECTOR_INITRAMFS_PATH" ]; then
+                if [ -e "$IRONIC_DEPLOY_KERNEL_PATH" -a -e "$IRONIC_DEPLOY_RAMDISK_PATH" ]; then
+                    cp $IRONIC_DEPLOY_KERNEL_PATH $IRONIC_INSPECTOR_KERNEL_PATH
+                    cp $IRONIC_DEPLOY_RAMDISK_PATH $IRONIC_INSPECTOR_INITRAMFS_PATH
+                else
+                    wget "$IRONIC_AGENT_KERNEL_URL" -O $IRONIC_INSPECTOR_KERNEL_PATH
+                    wget "$IRONIC_AGENT_RAMDISK_URL" -O $IRONIC_INSPECTOR_INITRAMFS_PATH
+                fi
+            fi
         fi
     else
-        IRONIC_INSPECTOR_KERNEL_PATH="$IRONIC_INSPECTOR_IMAGE_PATH.kernel"
         IRONIC_INSPECTOR_KERNEL_CMDLINE="discoverd_callback_url=$IRONIC_INSPECTOR_CALLBACK_URI inspector_callback_url=$IRONIC_INSPECTOR_CALLBACK_URI"
         if [ ! -e "$IRONIC_INSPECTOR_KERNEL_PATH" -o ! -e "$IRONIC_INSPECTOR_INITRAMFS_PATH" ]; then
             if [[ $(type -P ramdisk-image-create) == "" ]]; then
@@ -131,11 +136,25 @@ function prepare_tftp {
         fi
     fi
 
-    mkdir_chown_stack "$IRONIC_TFTPBOOT_DIR/pxelinux.cfg"
-    cp $IRONIC_INSPECTOR_KERNEL_PATH $IRONIC_TFTPBOOT_DIR/ironic-inspector.kernel
-    cp $IRONIC_INSPECTOR_INITRAMFS_PATH $IRONIC_TFTPBOOT_DIR
+    if [[ "$IRONIC_IPXE_ENABLED" == "True" ]] ; then
+        cp $IRONIC_INSPECTOR_KERNEL_PATH $IRONIC_HTTP_DIR/ironic-inspector.kernel
+        cp $IRONIC_INSPECTOR_INITRAMFS_PATH $IRONIC_HTTP_DIR
 
-    cat > "$IRONIC_TFTPBOOT_DIR/pxelinux.cfg/default" <<EOF
+        cat > "$IRONIC_HTTP_DIR/ironic-inspector.ipxe" <<EOF
+#!ipxe
+
+dhcp
+
+kernel http://$IRONIC_HTTP_SERVER:$IRONIC_HTTP_PORT/ironic-inspector.kernel BOOTIF=\${mac} $IRONIC_INSPECTOR_KERNEL_CMDLINE
+initrd http://$IRONIC_HTTP_SERVER:$IRONIC_HTTP_PORT/ironic-inspector.initramfs
+boot
+EOF
+    else
+        mkdir_chown_stack "$IRONIC_TFTPBOOT_DIR/pxelinux.cfg"
+        cp $IRONIC_INSPECTOR_KERNEL_PATH $IRONIC_TFTPBOOT_DIR/ironic-inspector.kernel
+        cp $IRONIC_INSPECTOR_INITRAMFS_PATH $IRONIC_TFTPBOOT_DIR
+
+        cat > "$IRONIC_TFTPBOOT_DIR/pxelinux.cfg/default" <<EOF
 default inspect
 
 label inspect
@@ -144,6 +163,7 @@ append initrd=ironic-inspector.initramfs $IRONIC_INSPECTOR_KERNEL_CMDLINE
 
 ipappend 3
 EOF
+    fi
 }
 
 function configure_inspector {
@@ -154,10 +174,10 @@ function configure_inspector {
 
     cp "$IRONIC_INSPECTOR_DIR/example.conf" "$IRONIC_INSPECTOR_CONF_FILE"
     inspector_iniset DEFAULT debug $IRONIC_INSPECTOR_DEBUG
-    inspector_iniset ironic os_auth_url "$KEYSTONE_SERVICE_URI/v2.0"
+    inspector_iniset ironic os_auth_url "$KEYSTONE_SERVICE_URI"
     inspector_iniset ironic os_username $IRONIC_INSPECTOR_ADMIN_USER
     inspector_iniset ironic os_password $SERVICE_PASSWORD
-    inspector_iniset ironic os_tenant_name $SERVICE_TENANT_NAME
+    inspector_iniset ironic os_tenant_name $SERVICE_PROJECT_NAME
 
     configure_auth_token_middleware $IRONIC_INSPECTOR_CONF_FILE $IRONIC_INSPECTOR_ADMIN_USER $IRONIC_INSPECTOR_AUTH_CACHE_DIR/api
 
@@ -195,13 +215,22 @@ function configure_inspector {
     inspector_iniset processing ramdisk_logs_dir "$IRONIC_INSPECTOR_RAMDISK_LOGDIR"
     inspector_iniset processing always_store_ramdisk_logs "$IRONIC_INSPECTOR_ALWAYS_STORE_RAMDISK_LOGS"
     inspector_iniset processing log_bmc_address False
+    if [ -n "$IRONIC_INSPECTOR_NODE_NOT_FOUND_HOOK" ]; then
+        inspector_iniset processing node_not_found_hook "$IRONIC_INSPECTOR_NODE_NOT_FOUND_HOOK"
+    fi
+    inspector_iniset DEFAULT timeout $IRONIC_INSPECTOR_TIMEOUT
+
+    get_or_create_service "ironic-inspector" "baremetal-introspection" "Ironic Inspector baremetal introspection service"
+    get_or_create_endpoint "baremetal-introspection" "$REGION_NAME" \
+        "$IRONIC_INSPECTOR_URI" "$IRONIC_INSPECTOR_URI" "$IRONIC_INSPECTOR_URI"
+
 }
 
 function configure_inspector_swift {
     inspector_iniset swift os_auth_url "$KEYSTONE_SERVICE_URI/v2.0"
     inspector_iniset swift username $IRONIC_INSPECTOR_ADMIN_USER
     inspector_iniset swift password $SERVICE_PASSWORD
-    inspector_iniset swift tenant_name $SERVICE_TENANT_NAME
+    inspector_iniset swift tenant_name $SERVICE_PROJECT_NAME
 
     inspector_iniset processing store_data swift
 }
@@ -209,14 +238,29 @@ function configure_inspector_swift {
 function configure_inspector_dhcp {
     mkdir_chown_stack "$IRONIC_INSPECTOR_CONF_DIR"
 
-    cat > "$IRONIC_INSPECTOR_DHCP_CONF_FILE" <<EOF
+    if [[ "$IRONIC_IPXE_ENABLED" == "True" ]] ; then
+        cat > "$IRONIC_INSPECTOR_DHCP_CONF_FILE" <<EOF
+no-daemon
+port=0
+interface=$IRONIC_INSPECTOR_INTERFACE
+bind-interfaces
+dhcp-range=$IRONIC_INSPECTOR_DHCP_RANGE
+dhcp-match=ipxe,175
+dhcp-boot=tag:!ipxe,undionly.kpxe
+dhcp-boot=tag:ipxe,http://$IRONIC_HTTP_SERVER:$IRONIC_HTTP_PORT/ironic-inspector.ipxe
+dhcp-sequential-ip
+EOF
+    else
+        cat > "$IRONIC_INSPECTOR_DHCP_CONF_FILE" <<EOF
 no-daemon
 port=0
 interface=$IRONIC_INSPECTOR_INTERFACE
 bind-interfaces
 dhcp-range=$IRONIC_INSPECTOR_DHCP_RANGE
 dhcp-boot=pxelinux.0
+dhcp-sequential-ip
 EOF
+    fi
 }
 
 function prepare_environment {
@@ -247,6 +291,9 @@ function create_ironic_inspector_cache_dir {
 function cleanup_inspector {
     rm -f $IRONIC_TFTPBOOT_DIR/pxelinux.cfg/default
     rm -f $IRONIC_TFTPBOOT_DIR/ironic-inspector.*
+    if [[ "$IRONIC_IPXE_ENABLED" == "True" ]] ; then
+        rm -f $IRONIC_HTTP_DIR/ironic-inspector.*
+    fi
     sudo rm -f /etc/sudoers.d/ironic-inspector-rootwrap
     sudo rm -rf $IRONIC_INSPECTOR_AUTH_CACHE_DIR
     sudo rm -rf "$IRONIC_INSPECTOR_RAMDISK_LOGDIR"
